Check if file already has been created

+++ Tools/createJava.py
@@ -7,7 +7,10 @@ for line in sys.stdin:
     print line
 
 fName = sys.argv[1]
-file = open(fName + ".java", 'w+')
+if(os.path.isfile(fName+".java")):
+    print "File Exists"
+else:
+    file = open(fName + ".java", 'w+')
 
 if (file):
     file.write("public class " + fName + " {\n    public static void main(String[] args) {\n        \n        \n    }\n}")
@@ -16,3 +19,6 @@ cmd = "javac " + fName +".java"
 
 subprocess.Popen(cmd, shell=True)
 
+cmd = "java " + fName
+
+subprocess.Popen(cmd, shell=True)
